split painter to small pieces.
